Спринт № 16, рефакторинг в MVVM... 9 (раскладываю по пакетам)
