--- app/src/main/java/com/example/playlistmaker/data/dto/TracksSearchResponse.kt
@@ -1,6 +1,0 @@
-package com.example.playlistmaker.data.dto
-
-class TracksSearchResponse(
-    val resultCount: Int,
-    val results: List<TrackDto>
-) : Response()
