+++ app/src/main/java/com/example/playlistmaker/data/search/NetworkClient.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker.data.search
+
+import com.example.playlistmaker.data.search.network.Response
+
+interface NetworkClient {
+    fun doRequest(dto: Any): Response
+
+}
