+++ app/src/main/java/com/example/playlistmaker/data/search/StorageClient.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.data
+package com.example.playlistmaker.data.search
 
 interface StorageClient<T> {
     fun storeData(data: T)
