+++ app/src/main/java/com/example/playlistmaker/data/search/dto/TrackDto.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.data.dto
+package com.example.playlistmaker.data.search.dto
 
 import java.text.SimpleDateFormat
 import java.util.Locale
