+++ app/src/main/java/com/example/playlistmaker/data/search/impl/SearchHistoryRepositoryImpl.kt
@@ -1,8 +1,8 @@
-package com.example.playlistmaker.data.storage
+package com.example.playlistmaker.data.search.impl
 
-import com.example.playlistmaker.data.StorageClient
+import com.example.playlistmaker.data.search.StorageClient
 import com.example.playlistmaker.domain.api.SearchHistoryRepository
-import com.example.playlistmaker.domain.models.Track
+import com.example.playlistmaker.domain.search.model.Track
 import com.example.playlistmaker.domain.util.Resource
 
 class SearchHistoryRepositoryImpl(
