+++ app/src/main/java/com/example/playlistmaker/data/search/network/Response.kt
@@ -1,0 +1,5 @@
+package com.example.playlistmaker.data.search.network
+
+open class Response() {
+    var resultCode = 0
+}
