+++ app/src/main/java/com/example/playlistmaker/data/search/network/RetrofitNetworkClient.kt
@@ -1,8 +1,6 @@
-package com.example.playlistmaker.data.network
+package com.example.playlistmaker.data.search.network
 
-import com.example.playlistmaker.data.NetworkClient
-import com.example.playlistmaker.data.dto.Response
-import com.example.playlistmaker.data.dto.TracksSearchRequest
+import com.example.playlistmaker.data.search.NetworkClient
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
