+++ app/src/main/java/com/example/playlistmaker/data/search/network/TracksSearchResponse.kt
@@ -1,0 +1,8 @@
+package com.example.playlistmaker.data.search.network
+
+import com.example.playlistmaker.data.search.dto.TrackDto
+
+class TracksSearchResponse(
+    val resultCount: Int,
+    val results: List<TrackDto>
+) : Response()
