--- app/src/main/java/com/example/playlistmaker/domain/api/DarkThemeRepository.kt
@@ -1,6 +1,0 @@
-package com.example.playlistmaker.domain.api
-
-interface DarkThemeRepository {
-    fun isDarkTheme(darkThemeDefaultValue: Boolean): Boolean
-    fun saveDarkTheme(darkTheme: Boolean)
-}
