+++ app/src/main/java/com/example/playlistmaker/domain/api/ExternalNavigator.kt
@@ -1,7 +1,7 @@
 package com.example.playlistmaker.domain.api
 
 import android.content.Context
-import com.example.playlistmaker.domain.models.EmailData
+import com.example.playlistmaker.domain.sharing.model.EmailData
 
 interface ExternalNavigator {
     fun shareLink(url: String, c: Context)
