+++ app/src/main/java/com/example/playlistmaker/domain/api/SearchHistoryInteractor.kt
@@ -1,6 +1,6 @@
 package com.example.playlistmaker.domain.api
 
-import com.example.playlistmaker.domain.models.Track
+import com.example.playlistmaker.domain.search.model.Track
 
 interface SearchHistoryInteractor {
