+++ app/src/main/java/com/example/playlistmaker/domain/api/SettingRepository.kt
@@ -1,6 +1,6 @@
 package com.example.playlistmaker.domain.api
 
-import com.example.playlistmaker.domain.models.ThemeSettings
+import com.example.playlistmaker.domain.settings.model.ThemeSettings
 
 interface SettingRepository {
     fun getThemeSettings(): ThemeSettings
