+++ app/src/main/java/com/example/playlistmaker/domain/api/SettingsRepository.kt
@@ -1,6 +1,6 @@
 package com.example.playlistmaker.domain.api
 
-import com.example.playlistmaker.domain.models.ThemeSettings
+import com.example.playlistmaker.domain.settings.model.ThemeSettings
 
 interface SettingsRepository {
     fun getThemeSettings(): ThemeSettings?
