+++ app/src/main/java/com/example/playlistmaker/domain/api/TracksInteractor.kt
@@ -1,7 +1,7 @@
 package com.example.playlistmaker.domain.api
 
 
-import com.example.playlistmaker.domain.models.Track
+import com.example.playlistmaker.domain.search.model.Track
 
 interface TracksInteractor {
     fun searchTracks(expression: String, consumer: TracksConsumer)
