+++ app/src/main/java/com/example/playlistmaker/domain/api/TracksRepository.kt
@@ -1,6 +1,6 @@
 package com.example.playlistmaker.domain.api
 
-import com.example.playlistmaker.domain.models.Track
+import com.example.playlistmaker.domain.search.model.Track
 import com.example.playlistmaker.domain.util.Resource
 
 interface TracksRepository {
