+++ app/src/main/java/com/example/playlistmaker/domain/impl/SearchHistoryInteractorImpl.kt
@@ -2,7 +2,7 @@ package com.example.playlistmaker.domain.impl
 
 import com.example.playlistmaker.domain.api.SearchHistoryInteractor
 import com.example.playlistmaker.domain.api.SearchHistoryRepository
-import com.example.playlistmaker.domain.models.Track
+import com.example.playlistmaker.domain.search.model.Track
 
 class SearchHistoryInteractorImpl(
     private val repository: SearchHistoryRepository
