+++ app/src/main/java/com/example/playlistmaker/domain/impl/SettingsInteractorImpl.kt
@@ -2,7 +2,7 @@ package com.example.playlistmaker.domain.impl
 
 import com.example.playlistmaker.domain.api.SettingsInteractor
 import com.example.playlistmaker.domain.api.SettingsRepository
-import com.example.playlistmaker.domain.models.ThemeSettings
+import com.example.playlistmaker.domain.settings.model.ThemeSettings
 
 class SettingsInteractorImpl(private val repository: SettingsRepository) : SettingsInteractor {
