--- app/src/main/java/com/example/playlistmaker/domain/models/ThemeSettings.kt
@@ -1,3 +1,0 @@
-package com.example.playlistmaker.domain.models
-
-data class ThemeSettings(var darkTheme: Boolean)
