+++ app/src/main/java/com/example/playlistmaker/domain/search/model/Track.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.domain.models
+package com.example.playlistmaker.domain.search.model
 
 data class Track(
     val trackId: Int,
