+++ app/src/main/java/com/example/playlistmaker/domain/settings/model/ThemeSettings.kt
@@ -1,0 +1,3 @@
+package com.example.playlistmaker.domain.settings.model
+
+data class ThemeSettings(var darkTheme: Boolean)
