+++ app/src/main/java/com/example/playlistmaker/ui/search/TrackAdapter.kt
@@ -4,7 +4,7 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.example.playlistmaker.R
-import com.example.playlistmaker.domain.models.Track
+import com.example.playlistmaker.domain.search.model.Track
 
 class TrackAdapter (private val clickListener: TrackClickListener) :
     RecyclerView.Adapter<TrackViewHolder> () {
