+++ app/src/main/java/com/example/playlistmaker/ui/search/TrackViewHolder.kt
@@ -11,7 +11,7 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
 import com.example.playlistmaker.R
 import com.example.playlistmaker.databinding.TrackItemBinding
-import com.example.playlistmaker.domain.models.Track
+import com.example.playlistmaker.domain.search.model.Track
 
 class TrackViewHolder(private val binding: TrackItemBinding): RecyclerView.ViewHolder(binding.root) {
